Play any music file by name that exists in the SD card using New SD lib.
Removed the Mp3SoftReset() from Mp3Reset(), as it will be executed every time playing new music file.
Another reason is musicshield will be stuck in the Mp3SoftReset() when power on, and hardware reset has to be pushed several times so that the board can run.

Next step is to remove usness files.